requirements.txt
streamlit>=1.28.0
ollama>=0.3.0
python-pptx>=0.6.21
PyMuPDF>=1.23.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
plotly>=5.18.0
